Apply the 'Libretto' theme.
The theme is from https://github.com/ferrolho/jekyll-theme-libretto

This theme doesn't appear to work as a gem, or as a target for the
Github Pages 'remote_theme' trick.  I do really like the theme's style,
so I've pulled in the content from theme into my site.  Hopefully this
doesn't screw things up too much on the Github side.